requirements.txt
# Django Core
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Database
psycopg2-binary==2.9.10

# Environment
python-dotenv==1.0.1
python-decouple==3.8

# Authentication
djangorestframework-simplejwt==5.4.0

# Admin Import/Export
django-import-export==4.3.4

# AI
openai==1.59.9

# Payments
razorpay==1.4.2

# Storage
boto3==1.36.4
django-storages==1.14.4

# Image Processing
Pillow==11.1.0

# WhatsApp API
requests==2.32.3

# Monitoring
sentry-sdk>=2.0.0

# Utilities
django-extensions==3.2.3
uuid6==2024.7.10

# WSGI
gunicorn==23.0.0

# Redis Cache
django-redis==5.4.0
redis==5.2.1

# Celery (Background Tasks)
celery==5.4.0

# API Documentation (Swagger/OpenAPI)
drf-spectacular==0.28.0

# HTMX Integration
django-htmx==1.21.0

# SMS/OTP (Twilio)
twilio==9.4.0
